product controller separate
