+++ src/main/java/com/lotteon/controller/controller/CartOrderController.java
@@ -1,0 +1,35 @@
+package com.lotteon.controller.controller;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/prod")
+@RequiredArgsConstructor
+@Log4j2
+public class CartOrderController {
+
+    @GetMapping("/cart")
+    public String join(Model model) {
+        return "pages/product/cart";
+    }
+
+    @GetMapping("/cart/direct")
+    public String cartDirect(Model model) {
+        return "redirect:/prod/order";
+    }
+
+    @GetMapping("/order")
+    public String order(Model model) {
+        return "pages/product/order";
+    }
+
+    @GetMapping("/order/complete")
+    public String orderComplete(Model model) {
+        return "pages/product/complete";
+    }
+}
